reduced CCN
